Revert "feat: add toggleable message/word chart"
